AyushiVashisth-07-03

+++ src/pages/AdminDashboard/Student.jsx
@@ -5,11 +5,10 @@ import {
   FaChevronLeft,
   FaChevronRight
 } from "react-icons/fa";
-import "../../styles/AdminDashboard/student.css";
+import "../../styles/AdminDashboard/Student.css";
 import Navbar from "../../components/AdminDashboard/Navbar";
 import StudentModal from "../../components/AdminDashboard/StudentModal";
-import { ToastContainer, toast } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
+import axios from "axios";
 
 function Student() {
   const [students, setStudents] = useState([]);
@@ -37,6 +36,7 @@ function Student() {
     )
       .then((response) => response.json())
       .then((data) => {
+        console.log("Student", data.Students);
         setStudents(data.Students);
       })
       .catch((error) => {
@@ -52,29 +52,30 @@ function Student() {
   };
 
   const deleteRow = (id) => {
-    setLoading(true);
-    console.log("id", id);
-    fetch(
-      `https://university-project-paresh.onrender.com/University/Admin/deleteStudent/${id}`,
-      {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json"
+    if (window.confirm(`Are you sure you want to delete this student?`)) {
+      setLoading(true);
+      console.log("id", id);
+      fetch(
+        `https://university-project-paresh.onrender.com/University/Admin/deleteStudent/${id}`,
+        {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json"
+          }
         }
-      }
-    )
-      .then((response) => response.json())
-      .then((response) => {
-        console.log(response);
-        toast.success(response.message);
-      })
-      .catch((error) => {
-        console.error("Error deleting student:", error);
-        toast.error("Failed to delete course. Please try again later.");
-      })
-      .finally(() => {
-        setLoading(false);
-      });
+      )
+        .then((response) => response.json())
+        .then((response) => {
+          console.log(response);
+          handleGetData();
+        })
+        .catch((error) => {
+          console.error("Error deleting student:", error);
+        })
+        .finally(() => {
+          setLoading(false);
+        });
+    }
   };
 
   const handlePaidClick = (rollNo) => {
@@ -82,10 +83,34 @@ function Student() {
     // You can perform API request here to mark the student as paid
   };
 
+  const handleDisableStudent = (id) => {
+    axios
+      .put(
+        `https://university-project-paresh.onrender.com/University/Admin/disableStudent/${id}`
+      )
+      .then((response) => {
+        handleGetData();
+      })
+      .catch((error) => {
+        console.error("Error disabling student:", error);
+      });
+  };
+
+  const handleEnableStudent = (id) => {
+    axios
+      .put(
+        `https://university-project-paresh.onrender.com/University/Admin/enableStudent/${id}`
+      )
+      .then((response) => {
+        handleGetData();
+      })
+      .catch((error) => {
+        console.error("Error enabling student:", error);
+      });
+  };
+
   const indexOfLastRow = currentPage * rowsPerPage;
   const indexOfFirstRow = indexOfLastRow - rowsPerPage;
-
-  // Filter students based on search query
   const filteredStudents = students.filter((student) =>
     Object.values(student).some((value) =>
       value.toString().toLowerCase().includes(searchQuery.toLowerCase())
@@ -103,7 +128,6 @@ function Student() {
       <div className="h-[60px] bg-black">
         <Navbar />
       </div>
-      <ToastContainer />
       <div className="bg-login bg-cover h-[89vh]">
         <section className="flex flex-col md:flex-row justify-between items-center p-2 rounded-lg mb-4 border-0 border-white mt-3">
           <h1 className="text-3xl font-semibold text-white text-center md:text-left mb-4 md:mb-0 md:mr-4 md:ml-0">
@@ -164,13 +188,19 @@ function Student() {
                     <th>Admission Year</th>
                     <th>Fees</th>
                     <th>Delete</th>
+                    <th>Disable</th>
+                    <th>Enable</th>
                   </tr>
                 </thead>
                 <tbody>
                   {currentRows.map((student, index) => (
                     <tr
                       key={index}
-                      className="bg-transparent"
+                      className={
+                        student.accountStatus === "disabled"
+                          ? "disabled-row"
+                          : ""
+                      }
                       onClick={() => handleRowClick(student)}
                     >
                       <td>{student.rollNo}</td>
@@ -200,6 +230,28 @@ function Student() {
                           className="border-0 text-red-700 rounded-md border-white"
                         >
                           <FaTrash className="mr-1" />
+                        </button>
+                      </td>
+                      <td>
+                        <button
+                          onClick={(e) => {
+                            e.stopPropagation();
+                            handleDisableStudent(student._id);
+                          }}
+                          className="border-0 text-red-700 rounded-md border-white"
+                        >
+                          Disable
+                        </button>
+                      </td>
+                      <td>
+                        <button
+                          onClick={(e) => {
+                            e.stopPropagation();
+                            handleEnableStudent(student._id);
+                          }}
+                          className="border-0 text-red-700 rounded-md border-white"
+                        >
+                          Enable
                         </button>
                       </td>
                     </tr>
@@ -233,7 +285,6 @@ function Student() {
           </button>
         </div>
       </div>
-      <ToastContainer />
     </div>
   );
 }
